fastapi-template/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.0.15
openai>=1.10.0,<2.0.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
chromadb==0.4.22
sentence-transformers==2.5.1
langchain==0.1.12
langchain-openai==0.0.8
